This version changed numerous variables from data type 'int' to 'double'.
**Successfully compiles BUT does not run due to 'parse_args()' error
(Configs error).

+++ agent.cpp
@@ -61,14 +61,14 @@ void Agent::initializeAll(int number, vector< vector<double> > agentValues)
     unitsSoldForDevicesToday = 0;
     unitsSoldCrossGroupForDevicesToday = 0;
 
-    unitsGatheredWithDeviceToday = vector< vector<int> >(NUM_DEVICE_TYPES);
+    unitsGatheredWithDeviceToday = vector< vector<double> >(NUM_DEVICE_TYPES);
 
-    devicesMadeWithDevDevicesToday = vector< vector<int> >(NUM_DEVICE_TYPES);
+    devicesMadeWithDevDevicesToday = vector< vector<double> >(NUM_DEVICE_TYPES);
 
     timeSpentMakingDevicesToday = vector<double>(NUM_DEVICE_TYPES, 0.0);
 
-    timeSpentGatheringWithDeviceToday = vector<int>(NUM_DEVICE_TYPES, 0.0);
-    timeSpentGatheringWithDeviceTodayByRes = vector< vector<int> >(NUM_DEVICE_TYPES);
+    timeSpentGatheringWithDeviceToday = vector<double>(NUM_DEVICE_TYPES, 0.0);
+    timeSpentGatheringWithDeviceTodayByRes = vector< vector<double> >(NUM_DEVICE_TYPES);
     timeSpentGatheringWithoutDeviceToday = 0.0;
 
     utilityToday = 0.0;
@@ -137,7 +137,7 @@ void Agent::defineProperties()
  * accessor method
  *  \return the number of units an agent holds of a given resource.
  */
-int Agent::getHeld(int resId) const
+double Agent::getHeld(int resId) const
 {
     return resProp[resId].getHeld();
 }
@@ -145,7 +145,7 @@ int Agent::getHeld(int resId) const
 /**
  * \return the number of units gathered today for resource resId
  */
-int Agent::getUnitsGatheredToday(int resId) const
+double Agent::getUnitsGatheredToday(int resId) const
 {
     return resProp[resId].unitsGatheredToday;
 }
@@ -153,7 +153,7 @@ int Agent::getUnitsGatheredToday(int resId) const
 /**
  * \return the number of units gahtered with device for resource resId
  */
-int Agent::getUnitsGatheredWithDevice(int device, int resId) const
+double Agent::getUnitsGatheredWithDevice(int device, int resId) const
 {
     return unitsGatheredWithDeviceToday[device][resId];
 }
@@ -161,7 +161,7 @@ int Agent::getUnitsGatheredWithDevice(int device, int resId) const
 /**
  * \return the time gathering resource with each device separately 
  */
-int Agent::getTimeSpentGatheringWithDeviceTodayByRes(int device, int resId) const
+double Agent::getTimeSpentGatheringWithDeviceTodayByRes(int device, int resId) const
 {
     return timeSpentGatheringWithDeviceTodayByRes[device][resId];
 }
@@ -169,7 +169,7 @@ int Agent::getTimeSpentGatheringWithDeviceTodayByRes(int device, int resId) cons
 /**
  * \return the time gathering resource by hand
  */
-int Agent::getTimeSpentGatheringWithoutDeviceTodayByRes(int resId) const
+double Agent::getTimeSpentGatheringWithoutDeviceTodayByRes(int resId) const
 {
     return resProp[resId].timeSpentGatheringWithoutDeviceToday;
 }
@@ -177,7 +177,7 @@ int Agent::getTimeSpentGatheringWithoutDeviceTodayByRes(int resId) const
 /**
  * \return the number of deviceIndex of deviceType made today.
  */
-int Agent::getDevicesMadeToday(int deviceIndex, int deviceType) const
+double Agent::getDevicesMadeToday(int deviceIndex, int deviceType) const
 {
     return devProp[deviceType][deviceIndex].devicesMadeToday;
 }
@@ -396,7 +396,7 @@ void Agent::deviceUse(device_name_t device, int deviceIndex, double timeChange)
  * required by the given resource bundle
  * \param bundle a vector of resource ids
  */
-bool Agent::resBundleHeldCheck(vector<int> bundle) const
+bool Agent::resBundleHeldCheck(vector<double> bundle) const
 {
     for (int resId = 0; resId < glob.NUM_RESOURCES; resId++) {
         if (bundle[resId] > resProp[resId].getHeld()) {
@@ -1425,7 +1425,7 @@ void Agent::sellsDevice(int deviceIndex, device_name_t device)
  * \param resourceBundle a vector of number of units of resource given up
  * \return the utility loss of giving up resourceBundle
  */
-double Agent::costOfResourceBundle(vector<int> &resourceBundle) const
+double Agent::costOfResourceBundle(vector<double> &resourceBundle) const
 {
     double cost = 0;
     for (int resId = 0; resId < glob.NUM_RESOURCES; resId++) {
@@ -1441,7 +1441,7 @@ double Agent::costOfResourceBundle(vector<int> &resourceBundle) const
  * \param resourceBundle a vector of number of units of resource gained
  * \return the utility gain of getting resourceBundle
  */
-double Agent::gainOfResourceBundle(vector<int> &resourceBundle) const
+double Agent::gainOfResourceBundle(vector<double> &resourceBundle) const
 {
     double gain = 0;
     for (int resId = 0; resId < glob.NUM_RESOURCES; resId++) {
@@ -1564,10 +1564,10 @@ MaxInfo Agent::calcMaxDeviceGainAndIndex(device_name_t deviceType)
          * is completely unable to buy any components).
          */
         if (devProp[deviceType][resId].getDeviceExperience() >= 1.0) {
-            pair < double, vector<int> > worstCaseConstruction =
+            pair < double, vector<double> > worstCaseConstruction =
                 consideredDevice[resId]->worstCaseConstruction(*this);
             double necessaryTime = worstCaseConstruction.first;
-            vector<int> necessaryRes = worstCaseConstruction.second;
+            vector<double> necessaryRes = worstCaseConstruction.second;
             if (resBundleHeldCheck(necessaryRes) &&
                 setAsideTime + necessaryTime < glob.DAY_LENGTH &&
                 deviceCurrentlyHeldForResource(resId, deviceType) < glob.DAYS_OF_DEVICE_TO_HOLD * glob.DAY_LENGTH) {
@@ -1748,7 +1748,7 @@ void Agent::resetDeviceGainAndCostMemory()
             thisDevProp.costOfDeviceMemory = 0.0;
             thisDevProp.costOfDeviceMemoryValid = false;
             thisDevProp.worstCaseConstructionMemory =
-                make_pair(0.0, vector<int>());
+                make_pair(0.0, vector<double>());
             thisDevProp.worstCaseConstructionMemoryValid = false;
         }
     }
@@ -1856,7 +1856,7 @@ void Agent::logAgentData()
     LOG(5) << "utilGainThroughDevSoldToday " << utilGainThroughDevSoldToday;
     LOG(5) << " --- resProps --- ";
     LOG(5) << "res     endDayUtil     held idleResource experience resSetAside unitsGath";
-    for (int i = 0; i < glob.NUM_RESOURCES; i++) {
+    for (double i = 0; i < glob.NUM_RESOURCES; i++) {
         char ostr[128];
         sprintf(ostr, "%2d\t%8.4f\t%3d\t%d\t%8.4f\t%4d\t%4d", i,
                 resProp[i].endDayUtilities, resProp[i].getHeld(),
@@ -1866,7 +1866,7 @@ void Agent::logAgentData()
     }
     LOG(5) << " --- devProps --- ";
     LOG(5) << "devExp  devHeld  idleDev devToMake devSetAsi gainOverDevLM Valid? costOfDevMem Valid? devMadeToday devMadeTot";
-    for (int i = 0; i < glob.NUM_RESOURCES; i++) {
+    for (double i = 0; i < glob.NUM_RESOURCES; i++) {
         for (int type = 0; type < NUM_DEVICE_TYPES; type++) {
             LOG(5) << " device " << i << ", type " << type;
             char ostr[128];
